+++ HISTORY.rst
@@ -22,6 +22,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * Add product validated field
+* Add product followers and product responsible role
 * Add sensefly invoice report
 * Add Delivery note / invoice report
 * Add team on manufacturing order